Change utf16+latin1 to always use alignment of 2
Currently the `realloc` signature doesn't allow for specifying both a
new and an old alignment so switching the alignment isn't valid for
existing implementations. Instead of starting with alignment 1 this
instead starts with alignment 2 for allocations related to latin1+utf16
strings and keeps the alignment at 2 even if the latin1 encoding ends up
being used.