Special-case empty `{ ... }` sequences for prettier name mangling.
With the current mangling for eg. records:

```python
  return 'record { ' + ', '.join(mangled_fields) + ' }'
```

the mangling for a record with 0 fields is `record {  }`, with two
spaces between the braces. That looks subjectively surprising.

This PR adds special cases to the mangling for `record` and similar
types so that the mangling is `record {}`, with no spaces between the
braces, which is subjectively prettier.

The argument against this is that empty records should be very rare,
so prettiness isn't important, and every special case makes the system
more complex. The argument for it is that a mangling with two spaces
looks weird.