Modernize the WIT.md description
This commit brings the `WIT.md` file up-to-date with recent developments
in WIT. I've added a new "introduction section" which is a higher-level
explanation of the format than the lexical/syntactic structure. I've
also edited the syntactic structure to clarify a few old references and
remove `resource`, `future`, and `stream` types for now. (resources to
be added soon I suspect)

Additionally I've added a description of a binary format for WIT based
on the component model. This binary format is intended to be the "types
only" mode of the `wasm-tools component` tooling, although it doesn't
match the current implementation and the `wasm-tools` repo will need to
be updated. This representation, however, provides the means by which to
understand what it means for a component to have the type of a `world`,
namely it's a subtype of the binary representation's component type.