Fix lifting an own handle from a borrow handle
This fixes what I believe is a mistake in the `CanonicalABI.md`
explainer currently where a lifting an owned resource handle from the
table can succeed if the entry in the table was actually a borrowed
handle. This inserts a check to ensure that the handle is an `own`
handle before completing the lifting operation.