Add more documentation about types. (#288)
Add documentation describing the floating-point types, and briefly
describe the difference between core-wasm numeric types and
component-level numeric types.

Add more subsection headings to the prose describing the types, breaking
up a sizeable quantity of text, and adding anchors that can be linked
to. In particular, I've often wanted an anchor for "handle types".

And, fix the number of definition types.