Add type definition for `rectype`
This changes the component model specification to reference a `rectype`
in its type defintions. This makes sense since the [GC] proposal has
graduated to become standard WebAssembly and the previous definitions
refer to `arraytype` and `structtype` which are subsumed by `rectype`.
Adding this also benefits the [shared-everything-threads] proposal,
which uses `shared` bits on composite types.

In talking with @alexcrichton about this, the `functype` alternative is
retained for now to allow backward compatibility for existing components
(e.g., components using the `0x60` prefix to define a core `functype`).
In the future, the `functype` alternative should be removed completely
(since it is subsumed under `rectype`). Potentially the `0x00` prefix
could be tweaked as well. In the meantime, this change allows more than
one way to encode a `functype`.

[GC]: https://github.com/WebAssembly/gc
[shared-everything-threads]: https://github.com/WebAssembly/shared-everything-threads

Co-authored-by: Alex Crichton <alex@alexcrichton.com>