Reorganize the "Canonical Built-ins" section of Explainer.md.
This moves the signatures for the canonical builtins out of prose paragraphs
and into tables, which highlights them and simplifies the prose.

This also adds "conceptual signatures". The canonical ABI signatures are
useful when one is writing compilers or bindings generators or other tools,
while the conceptual signatures are useful for people seeking to
understand what these builtins do at a conceptual level, before thinking
about how all the various values are lowered into `i32`s.

And, the conceptual signatures in theory are independent of the ABI, so
they could be reused when new ABIs are added.

This is a follow-up to WebAssembly/component-model#407; I've re-added the
`canon` ebpf for now, removed the "Binary encoding immediates" table rows,
and tidied up the signatures to avoid using `...`.