Gate some spec features separately from async
This commit comes out of some discussion with folks today where a few
features of the async specification are going to be gated behind
separate feature gates distinct from the current one. This is to help
reduce scope slightly for the 0.3.0 release of WASI and focus on the
other features to ship. Here two new emoji-features are added:

* 🚝: marking some builtins as `async`
* 🚟: using `async` with `canon lift` without `callback` (stackful abi)

I've tried to update `Async.md` and `CanonicalABI.md` where appropriate
with these new emoji.