Document some filesystem conventions of WIT
I've been meaning to write these down for awhile, and now I've gotten
around to them! The goal is to outline some preexisting conventions for
specifying WIT to bindings generators in a location that isn't buried in
any implementation but instead in a shared location that can be iterated
on as well.