clarify canonopt requirements for async lifts
I had _thought_ that the spec allowed an async lift to omit the memory canonopt
if only the result type (and not the paarameter types) required it, leaving the
core function free to call any `task.return` at runtime as long as the type and
string encoding matched.  Indeed, that's what we implemented in `wasm-tools` and
Wasmtime.

However, it looks like I misremembered, since the spec uses an exact comparison
in `LiftOptions.equal`.  That's good news, because allowing the core function to
call any of several `task.return` functions, each with a different memory, makes
it hard for the runtime to optimize compositions.

So no actual change needed here -- just a clarification.  We'll update
`wasm-tools` and `Wasmtime` to match.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>